requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.13
-e git+https://github.com/fastapi/fastapi.git@53d2453d1a77f3384a1648d717f8ddafb5e9e460#egg=fastapi
fonttools==4.61.1
h11==0.16.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
murmurhash==1.0.15
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
preshed==3.0.12
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
requests==2.32.5
six==1.17.0
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.50.0
thinc==8.3.10
tqdm==4.67.1
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1
